enclave-kms/requirements.txt
eth-account>=0.8.0
python-dotenv >= 0.19.0
Flask==3.1.0
Flask-Cors==5.0.0